Client and C++ universe are exchanging BSON messages now.
Also cleaned up the BSON API, using function signature overloading instead of different functions. Less explicit to read, but nicer to write.

+++ Clients/Diana2DClient/Diana2DClient/frmHelm.cs
@@ -19,7 +19,7 @@ namespace Diana2DClient
         Smarty smarty = null;
         TcpClient client = null;
         Stream s = null;
-        int clientID = -1;
+        Int64 clientID = -1;
 
         Bitmap needle;
         Graphics needleG;
